requirements.txt
torch==2.4.1
torchvision==0.15.2
streamlit==1.19.0
numpy>=1.24.0
Pillow>=9.0.0